tools: mixins.py -> abc.py
