+++ devito/tools/__init__.py
@@ -1,8 +1,8 @@
 from devito.tools.utils import *  # noqa
 
+from devito.tools.abc import *  # noqa
 from devito.tools.algorithms import *  # noqa
 from devito.tools.data_structures import *  # noqa
 from devito.tools.memoization import *  # noqa
-from devito.tools.mixins import *  # noqa
 from devito.tools.validators import *  # noqa
 from devito.tools.visitors import *  # noqa
